The next iteration of the test code for min_count_autoadjust.
If I make the word vectors large enough that memory is exhausted
quickly, the kernel out-of-memory manager kills it before it has a
chance to adjust the vocabulary. If I make the word vectors smaller,
it has to make the sentence lists too large. I ran it for 12 hours
without it running the final min_count_autoadjust iteration. I just
realized that my sentence iterator is just returning large single-word
sentences. Oops. I'll try fixing that and see if things improve.